fix(bot): update Junetic.Common package to 0.4.0

+++ src/PasswordManager.Bot/Services/Abstractions/ITelegramBotUi.cs
@@ -37,7 +37,7 @@ public interface ITelegramBotUi {
 	/// <param name="pageIndex"></param>
 	/// <param name="searchQuery">query of account search (basically name of account we need to find) used to place it in button for future searches if we have more than 1 page</param>
 	/// <param name="messageToEditId">Used to edit same message when navigating through search result pages</param>
-	Task ShowAccountsPageAsync(BotUser botUser, IReadOnlyList<Account> accounts, int totalAccountCount, int pageIndex, string searchQuery, int messageToEditId = 0);
+	Task ShowAccountsPageAsync(BotUser botUser, IReadOnlyList<Account> accounts, long totalAccountCount, int pageIndex, string searchQuery, int messageToEditId = 0);
 
 	/// <summary>
 	/// Serializes given account to MarkdownV2 string
